Remove language extention from remaining binaries.
The install target install some binaries without language extentions and
others with such extension:

 * grgsm_capture.py
 * grgsm_channelize.py
 * grgsm_decode
 * grgsm_livemon
 * grgsm_scanner

It would be more consistent and according to the policy of Debian and
Ubuntu if all the binaries were without such language extention.  This
commit rename grgsm_capture.py and grgsm_channelize.py to avoid the
language extention.