Connect Message Printer block to UDP server block.
Instead of printing messages in parallel to the UDP package sending,
move the printing to after the UDP server block.  This make it both
possible to disable message printing by changing the server port to a
different value from the client port, and to print messages collected
by several clients by pointing all clients to the same server.