Return early on error to match the previous behavior (#353)
Removed unnecessary return statement at the end of the function. The return type is void.
Some bursts are different before and after the said commit in #353. This doesn't seem to affect the resulting data, as they seem to be corrupt anyway.
Note: There are updates to "lib/decoding/osmocom/coding/" at the upstream. I don't know if they're needed for something, though.

Happy holidays!

Signed-off-by: Christian Inci <chris.gh@broke-the-inter.net>